+++ history.md
@@ -1,4 +1,12 @@
 # 修改历史
+## 0.8.0 (2020-02-23)
+* feat: afterTask 相关函数已迁移到 `yyl-seed-gulp-requirejs` 本组件不再保留
+* del: 移除 `yh.optimize.afterTask()`
+* del: 移除 `yh.optimize.varSugar()`
+* del: 移除 `yh.optimize.resource()`
+* del: 移除 `yh.optimize.concat()`
+* del: 移除 `yh.optimize.rev`
+
 ## 0.7.1(2020-02-23)
 * feat: 新增 `yh.optimize.getHomePage()` 方法